Get rid of ENV hacking and fix the baseURL setting.
Add `getAssetURL` as a helper in the `RenderApp::Plugin::Assets` plugin
module. This method should be used for any javascript or css files to
ensure that the baseURL setting is correctly utilized.

The language used for determining if a right-to-left css file is
needed is an optional last argument.  When rendering the `language` form
parameter can set this, and in all cases the fall back is the `language`
set in the app configuration.

The only environment variables that are set are those that are needed
for rendering.  The renderer code itself should generally pass the
controller object in where needed and configuration values should be
extracted from there instead of setting environment variables
corresponding to each configuration value.

Also move the public/js/apps directories directly in public/js.  This
was done for webwork2 and PG a while ago, and should be done here as
well.